refactor: Chagne file path of Elements
