+++ src/js/components/EventLoop.js
@@ -1,5 +1,5 @@
 import { ANIMATION } from "../utils/Constants.js";
-import Elements from "../utils/Elements.js";
+import Elements from "./Elements.js";
 
 async function excuteEventLoop(memories) {
 	//TODO refactor
