+++ src/js/components/MicroQueue.js
@@ -1,5 +1,5 @@
 import { ANIMATION } from "../utils/Constants.js";
-import Elements from "../utils/Elements.js";
+import Elements from "./Elements.js";
 
 class MicroQueue {
 	#block;
