+++ src/js/main.js
@@ -1,4 +1,4 @@
-import Elements from "./utils/Elements.js";
+import Elements from "./components/Elements.js";
 import addEvnetCodeExcute from "./view/code.js";
 
 Elements.init();
